+++ packages/core/src/index.ts
@@ -1,7 +1,6 @@
 export * from './constants';
 export * from './core';
 export * from './types';
-export * from './utils';
 
 import { ETransferCore } from './core';
 export const eTransferCore = new ETransferCore({});