+++ packages/services/src/services.ts
@@ -7,23 +7,23 @@ import {
   TCreateWithdrawOrderResult,
   TGetAuthRequest,
   TGetAuthResult,
-  TGetDepositCalculateRequest,
-  TGetDepositCalculateResult,
-  TGetDepositInfoRequest,
-  TGetDepositInfoResult,
-  TGetNetworkListRequest,
-  TGetNetworkListResult,
-  TGetRecordStatusResult,
-  TGetRecordsListRequest,
-  TGetRecordsListResult,
   TGetTokenListRequest,
   TGetTokenListResult,
   TGetTokenOptionRequest,
   TGetTokenOptionResult,
+  TGetNetworkListRequest,
+  TGetNetworkListResult,
+  TGetDepositInfoRequest,
+  TGetDepositInfoResult,
+  TGetDepositCalculateRequest,
+  TGetDepositCalculateResult,
   TGetWithdrawInfoRequest,
   TGetWithdrawInfoResult,
-  TServices,
-} from './types';
+  TGetRecordsListRequest,
+  TGetRecordsListResult,
+  TGetRecordStatusResult,
+} from '@etransfer/types';
+import { TServices } from './types';
 import { formatApiError } from './utils';
 import { API_LIST, AUTH_API_BASE_PARAMS, CancelTokenSourceKey } from './constants';
 import { TRequestConfig, EtransferRequest } from '@etransfer/request';