+++ packages/types/src/index.ts
@@ -1,2 +1,3 @@
 export * from './portkey';
 export * from './services';
+export * from './storage';